cast to double for std::format
with newer minor version of VS19 the
std::format does not compile anymore
with int8_t input so cast to double

+++ src/dbg_gui_windows.cpp
@@ -11,7 +11,7 @@
 
 template <typename T>
 inline std::string numberAsStr(T number) {
-    return std::format("{:g}", number);
+    return std::format("{:g}", double(number));
 }
 
 inline static std::string getSourceValueStr(ValueSource src) {
